+++ Apps/Contoso.UWP.Puppet/MainPage.xaml.cs
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-using System;
 using Microsoft.AppCenter.Analytics;
+using System;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;